backend/ml-service/requirements.txt
tensorflow>=2.16.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
psycopg2-binary>=2.9.0
transformers>=4.35.0
torch>=2.1.0

